refactor: complete userspace manager split
